Seperating Animations from Local Physics
Originally followed a tutorial where the physics was tied to the animations. Led to too many bugs so I untangled the systems and essentially started to jump from near scratch. Additionally added the LocalPhysicsEngine to simplify the Controller element.

Collision is still broken. Now it can be worked on as animation has been untangled.